[장준영] 회원가입시 사진 미리보기 크기 조절 / 파일 업로드 및 DB저장+이메일인증API+중복검사 진행중

+++ booking_c/src/main/java/goodee/gdj58/booking_c/vo/CustomerImg.java
@@ -8,5 +8,5 @@ public class CustomerImg {
 	private String customerImgSaveName;
 	private String customerImgOriginName;
 	private String customerImgKind;
-	private int customerImgSize;
+	private long customerImgSize;
 }
